refactor: unit test of just input
